+++ docs/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.83.0
+
+### Added
+
+- [rule/dependency](checks/rule/dependency.md) check now supports `ignoreGroupMismatch`
+  option for ignoring specific metrics when checking for cross-group dependencies
+  - [#1795](https://github.com/cloudflare/pint/issues/1795).
+
 ## v0.82.3
 
 ### Fixed